Upgrade base docker image cuda 7.5 with ubuntu 14 to cuda 8.0 with ubuntu 16
cmake will fail on ubuntu 14, so we need to move ubuntu 16. (and cuda 7.5 with ubuntu 16 does not exist)

And also python-tk, pip upgrade and pip setuptools are needed during cmake.

That ubuntu image does not have the major editor (e.g. vim, ed, nano, emacs and so on) and `apt-get install vim` in the container does not work, so I add it for avoiding the inconvinience.